move decode.go file to util pkg
